Rename and move delimited ZSink.
